Remove unnecessary view update
