+++ src/app/App.svelte
@@ -1,10 +1,8 @@
 <script lang="ts">
-  import { Callout, Loading, Typography } from "obsidian-svelte";
   import { onMount } from "svelte";
 
   import { createProject } from "src/lib/data-api";
   import { api } from "src/lib/stores/api";
-  import { dataFrame, dataSource } from "src/lib/stores/dataframe";
   import { i18n } from "src/lib/stores/i18n";
   import { app } from "src/lib/stores/obsidian";
   import { settings } from "src/lib/stores/settings";
@@ -20,18 +18,6 @@
   let viewId: string | undefined;
 
   $: ({ projects } = $settings);
-
-  let querying: Promise<void>;
-
-  $: {
-    // Perform a full refresh of the data frame whenever the data source
-    // changes.
-    querying = (async () => {
-      if ($dataSource) {
-        dataFrame.set(await $dataSource.queryAll());
-      }
-    })();
-  }
 
   onMount(() => {
     if (!projects.length) {
@@ -54,8 +40,6 @@
       ).open();
     }
   });
-
-  const wait = () => new Promise((res) => setTimeout(res, 500));
 </script>
 
 <!--
@@ -64,27 +48,23 @@
 	App is the main application component and coordinates between the View and
 	the Toolbar.
 -->
-<AppContainer {projects} bind:projectId bind:viewId let:project let:view>
-  {#await querying}
-    {#await wait() then}
-      <Loading />
-    {/await}
-  {:then}
-    {#if project && view && $dataSource}
-      <View
-        {project}
-        {view}
-        readonly={$dataSource.readonly()}
-        api={new ViewApi($app, $dataSource, $api)}
-        onConfigChange={settings.updateViewConfig}
-        frame={$dataFrame}
-      />
-    {/if}
-  {:catch error}
-    <div style="padding: var(--size-4-3)">
-      <Callout title={error.name} icon="zap" variant="danger">
-        <Typography variant="body">{error.message}</Typography>
-      </Callout>
-    </div>
-  {/await}
+<AppContainer
+  {projects}
+  bind:projectId
+  bind:viewId
+  let:project
+  let:view
+  let:source
+  let:frame
+>
+  {#if project && view && source}
+    <View
+      {project}
+      {view}
+      readonly={source.readonly()}
+      api={new ViewApi($app, source, $api)}
+      onConfigChange={settings.updateViewConfig}
+      {frame}
+    />
+  {/if}
 </AppContainer>
