+++ src/app/AppContainer.svelte
@@ -1,10 +1,11 @@
 <script lang="ts">
   import type { App } from "obsidian";
+  import { Callout, Loading, Typography } from "obsidian-svelte";
 
   import type { DataSource } from "src/lib/data";
   import { DataviewDataSource } from "src/lib/datasources/dataview/dataview";
   import { FrontMatterDataSource } from "src/lib/datasources/frontmatter/frontmatter";
-  import { dataSource } from "src/lib/stores/dataframe";
+  import { dataFrame, dataSource } from "src/lib/stores/dataframe";
   import { app } from "src/lib/stores/obsidian";
   import type { ProjectDefinition } from "src/types";
 
@@ -32,6 +33,18 @@
     }
   }
 
+  let querying: Promise<void>;
+
+  $: {
+    // Perform a full refresh of the data frame whenever the data source
+    // changes.
+    querying = (async () => {
+      if ($dataSource) {
+        dataFrame.set(await $dataSource.queryAll());
+      }
+    })();
+  }
+
   // resolveDataSource selects the data source to use based on the project
   // settings.
   function resolveDataSource(project: ProjectDefinition, app: App): DataSource {
@@ -39,6 +52,8 @@
       ? new DataviewDataSource(app, project)
       : new FrontMatterDataSource(app, project);
   }
+
+  const wait = () => new Promise((res) => setTimeout(res, 500));
 </script>
 
 <!--
@@ -56,7 +71,24 @@
   />
 
   <div class="projects-main">
-    <slot project={selectedProject} view={selectedView} />
+    {#await querying}
+      {#await wait() then}
+        <Loading />
+      {/await}
+    {:then}
+      <slot
+        project={selectedProject}
+        view={selectedView}
+        source={$dataSource}
+        frame={$dataFrame}
+      />
+    {:catch error}
+      <div style="padding: var(--size-4-3)">
+        <Callout title={error.name} icon="zap" variant="danger">
+          <Typography variant="body">{error.message}</Typography>
+        </Callout>
+      </div>
+    {/await}
   </div>
 </div>
